refactor: CommentType.ts 内で定義の type の Suffix を type に統一
および、review.tsx でしか使用されていない Comment の定義をファイル内へ移動
