--- src/types/CommentType.ts
@@ -1,11 +1,0 @@
-export type Comment = {
-  path: string;
-  line: number;
-  side: string;
-  body: string;
-};
-
-export type PreviewComment = {
-  path: string;
-  body: string;
-};
