+++ src/types/PreviewCommentType.ts
@@ -1,0 +1,4 @@
+export type PreviewComment = {
+  path: string;
+  body: string;
+};
